refactor: move FilterMenuTitle to helpers file
