+++ packages/shared/src/components/filters/helpers.ts
@@ -3,6 +3,15 @@ import {
   AdvancedSettingsGroup,
 } from '../../graphql/feedSettings';
 
+export enum FilterMenuTitle {
+  MyFeed = 'My feed',
+  Tags = 'Manage tags',
+  ManageCategories = 'Manage categories',
+  ContentTypes = 'Content & Language',
+  Blocked = 'Blocking',
+  Following = 'Following',
+}
+
 export const getContentFromGroup = (
   advancedSettings: AdvancedSettings[],
   groupProp: AdvancedSettingsGroup[],
